third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-sandboxed-iframe-denied-without-wildcard-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/access-control-sandboxed-iframe-denied-without-wildcard.cgi. The 'Access-Control-Allow-Origin' whitelists only 'http://127.0.0.1:8000'. Origin 'null' is not in the list, and is therefore not allowed access.
This test verifies that sandboxed iframe does not have XmlHttpRequest access to its server with "Access-Control-Allow-Origin" set to its own origin (127.0.0.1).

This test will print "PASS" on success.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: Exception thrown. Sandboxed iframe XHR access was denied in 'send'. [A network error occurred.].
